requirements.txt
absl-py==0.12.0
anyio==3.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.3
colorama==0.4.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.0
dnspython==2.1.0
dominate==2.6.0
entrypoints==0.3
filelock==3.0.12
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-WTF==1.0.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.1
Jinja2==3.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
Keras-Preprocessing==1.1.2
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==2.0.1
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandocfilters==1.4.3
parso==0.7.0
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.5
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pymongo==3.12.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-bsonjs==0.2.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2021.1
pywin32==228
pywinpty==1.1.2
pyzmq==19.0.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.4
visitor==0.1.3
wcwidth==0.2.4
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.2
wrapt==1.12.1
WTForms==3.0.0
zipp==3.1.0
